+++ app/helpers/aspect_global_helper.rb
@@ -17,14 +17,12 @@ module AspectGlobalHelper
     if stream
       aspects = stream.aspects
       aspect = stream.aspect
-      aspect_ids = stream.aspect_ids
     elsif current_user
       aspects = current_user.post_default_aspects
       aspect = aspects.first
-      aspect_ids = current_user.aspect_ids
     else
       return {}
     end
-    {selected_aspects: aspects, aspect: aspect, aspect_ids: aspect_ids}
+    {selected_aspects: aspects, aspect: aspect}
   end
 end